Speed up and slow down multipliers added as constants
git-svn-id: https://ophelia.cs.put.poznan.pl/svn/prv-jdyna@46 7041bab8-1725-11de-84ca-a9d339809633

+++ src/java/org/jdyna/PlayerInfo.java
@@ -32,7 +32,7 @@ final class PlayerInfo implements IPlayerSprite
      * If player collects the speed bonus the variable is changed. 
      * If player doesn't have speed bonus the variable is set to 1.0. 
      */
-    double speedMultiplier = 1.0;
+    float speedMultiplier = 1.0f;
     
     /**
      * An increasing counter of frames if the player is in walking state.
